requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
hashids==1.3.1
pymysql
